`tools/importer-rest-api-specs`: refactoring the transforms for the repository models out into it's own package
This both makes the migration to the new models simpler (since it's contained within one package) but also means the parent package can be tidied per Phase 10
