+++ tools/importer-rest-api-specs/components/dataapigeneratorjson/generate_service.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 	"path"
 
+	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/dataapigeneratorjson/transforms"
 	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/models"
 )
 
@@ -15,7 +16,7 @@ func (s Generator) generateServiceDefinitions(apiVersions []models.AzureApiDefin
 	s.logger.Debug(fmt.Sprintf("Processing Service %q..", s.serviceName))
 
 	s.logger.Trace(fmt.Sprintf("Generating Service Definition Code for Service %q..", s.serviceName))
-	serviceDefinition, err := buildServiceDefinition(s.serviceName, s.resourceProvider, s.terraformPackageName, apiVersions)
+	serviceDefinition, err := transforms.MapServiceDefinitionToRepository(s.serviceName, s.resourceProvider, s.terraformPackageName, apiVersions)
 	if err != nil {
 		return fmt.Errorf("mapping Service Definition for %q: %+v", s.serviceName, err)
 	}
